+++ packages/material/src/MaterialShiftActions.tsx
@@ -3,16 +3,15 @@ import * as React from 'react';
 import type { ShiftActionsProps } from 'react-querybuilder';
 import { ShiftActions } from 'react-querybuilder';
 import { RQBMaterialContext } from './RQBMaterialContext';
-import type { RQBMaterialComponents } from './types';
+import type { MuiAugmentation } from './types';
 
 /**
  * @group Props
  */
 export interface MaterialShiftActionsProps
   extends ShiftActionsProps,
-    React.ComponentPropsWithoutRef<typeof Button> {
-  muiComponents?: RQBMaterialComponents;
-}
+    React.ComponentPropsWithoutRef<typeof Button>,
+    MuiAugmentation {}
 
 /**
  * @group Components