+++ packages/material/src/utils.tsx
@@ -8,9 +8,17 @@ export { isOptionGroupArray };
 
 type ToOptionsOptions = Pick<RQBMaterialComponents, 'ListSubheader' | 'MenuItem'>;
 
+// istanbul ignore next
+const defaultToOptionsOptions: ToOptionsOptions = {
+  ListSubheader: () => null,
+  MenuItem: () => <></>,
+};
+
 export const toOptions = (
-  arr: OptionList,
-  { ListSubheader, MenuItem }: ToOptionsOptions
+  // istanbul ignore next
+  arr: OptionList = [],
+  // istanbul ignore next
+  { ListSubheader, MenuItem }: ToOptionsOptions = defaultToOptionsOptions
 ): JSX.Element[] | null => {
   if (isOptionGroupArray(arr)) {
     const optArray: JSX.Element[] = [];